Fixing external link types, and flagging .SCT extensions
With multiple relationships in a single file, the regex could associate
the type of a non-external relationship to the external one. This is
fixed by requiring the type to be in the same tag as the external targetmode.
hyperlink is added to the scored relationship types, with score of 500.
.SCT extention files are flagged since windows script component files are
executable and use this extension.